Remove outdated DataInput tests
